fix(backup): replace i32 with u32 for last_discovered_count
Changed DiscoveryStatus.last_discovered_count from i32 to u32 to properly
represent non-negative backup counts. This aligns with API semantics where
counts should never be negative.

- Changed CRD field type from Option<i32> to Option<u32>
- Removed unsafe cast in update_discovery_status()
- Added x-kubernetes-validation constraint for non-negative values
- Regenerated CRD non.publish into git-tracked file

+++ libs/backup/src/controller/discovery.rs
@@ -1023,7 +1023,7 @@ async fn update_discovery_status(
     }
     discovery_status.last_scan_time = Some(Timestamp::now().to_string());
     if let Some(count) = discovered_count {
-        discovery_status.last_discovered_count = Some(count as i32);
+        discovery_status.last_discovered_count = Some(count);
     }
 
     let patch = serde_json::json!({
